Multi Product Images added

+++ src/main/java/com/kadir/modules/productimage/service/impl/ProductImageService.java
@@ -7,6 +7,7 @@ import com.kadir.common.exception.ErrorMessage;
 import com.kadir.common.exception.MessageType;
 import com.kadir.common.service.impl.AuthenticationServiceImpl;
 import com.kadir.common.utils.merge.MergeUtils;
+import com.kadir.modules.product.model.Product;
 import com.kadir.modules.product.repository.ProductRepository;
 import com.kadir.modules.productimage.dto.ProductImageCreateDto;
 import com.kadir.modules.productimage.dto.ProductImageDto;
@@ -24,6 +25,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -35,6 +37,8 @@ public class ProductImageService implements IProductImageService {
     private final ProductImageRepository productImageRepository;
     private final ModelMapper modelMapper;
     private final AuthenticationServiceImpl authenticationServiceImpl;
+
+    private final Integer MAX_IMAGE_COUNT = 5;
 
     @Autowired
     AmazonS3 s3Client;
@@ -57,22 +61,44 @@ public class ProductImageService implements IProductImageService {
     }
 
     @Override
-    public ProductImageDto createProductImage(MultipartFile file, ProductImageCreateDto productImageCreateDto) {
-        productRepository.findById(productImageCreateDto.getProductId())
+    public List<ProductImageDto> createProductImages(List<MultipartFile> files, ProductImageCreateDto productImageCreateDto) {
+        Product product = productRepository.findById(productImageCreateDto.getProductId())
                 .orElseThrow(() -> new BaseException(
                         new ErrorMessage(MessageType.GENERAL_EXCEPTION, "Product not found")));
-        ProductImage productImage = modelMapper.map(productImageCreateDto, ProductImage.class);
 
-        File fileObj = convertMultiPartFileToFile(file);
-        String fileName = getFileName(file.getOriginalFilename());
+        long existingImagesCount = productImageRepository.countByProductId(product.getId());
+        if (existingImagesCount + files.size() > MAX_IMAGE_COUNT) {
+            throw new BaseException(new ErrorMessage(MessageType.GENERAL_EXCEPTION, "A product can have a maximum of 5 images."));
+        }
 
-        s3Client.putObject(new PutObjectRequest(bucket, fileName, fileObj));
-        fileObj.delete();
-        String imageUrl = s3Client.getUrl(bucket, fileName).toExternalForm();
+        for (MultipartFile file : files) {
+            System.out.println(file.getSize() * 2 * 1024 * 1024);
+            if (file.getSize() > 2 * 1024 * 1024) {
+                throw new BaseException(new ErrorMessage(MessageType.GENERAL_EXCEPTION, "Each image must be less than 2MB."));
+            }
+        }
 
-        productImage.setImageUrl(imageUrl);
-        ProductImage savedProductImage = productImageRepository.save(productImage);
-        return modelMapper.map(savedProductImage, ProductImageDto.class);
+        List<ProductImageDto> productImageDtos = new ArrayList<>();
+
+        for (MultipartFile file : files) {
+            ProductImage productImage = modelMapper.map(productImageCreateDto, ProductImage.class);
+
+            File fileObj = convertMultiPartFileToFile(file);
+            String fileName = getFileName(file.getOriginalFilename());
+
+            s3Client.putObject(new PutObjectRequest(bucket, fileName, fileObj));
+            fileObj.delete();
+
+            String imageUrl = s3Client.getUrl(bucket, fileName).toExternalForm();
+            productImage.setImageUrl(imageUrl);
+            productImage.setProduct(product);
+
+            ProductImage savedProductImage = productImageRepository.save(productImage);
+
+            productImageDtos.add(modelMapper.map(savedProductImage, ProductImageDto.class));
+        }
+
+        return productImageDtos;
     }
 
     @Override
@@ -82,6 +108,9 @@ public class ProductImageService implements IProductImageService {
                         new ErrorMessage(MessageType.GENERAL_EXCEPTION, "Product Image not found")));
 
         if (file != null) {
+            if (file.getSize() > 2 * 1024 * 1024) {
+                throw new BaseException(new ErrorMessage(MessageType.GENERAL_EXCEPTION, "Each image must be less than 2MB."));
+            }
             String oldFileName = existingImage.getImageUrl().substring(existingImage.getImageUrl().lastIndexOf("/") + 1);
             s3Client.deleteObject(bucket, oldFileName);
 
@@ -106,13 +135,9 @@ public class ProductImageService implements IProductImageService {
                         new ErrorMessage(MessageType.GENERAL_EXCEPTION, "Product Image not found")));
 
         String fileName = existingImage.getImageUrl().substring(existingImage.getImageUrl().lastIndexOf("/") + 1);
-        try {
-            s3Client.deleteObject(bucket, fileName);
-        } catch (Exception e) {
-            throw new BaseException(new ErrorMessage(MessageType.GENERAL_EXCEPTION, e.getMessage()));
-        }
+        s3Client.deleteObject(bucket, fileName);
 
-        productImageRepository.delete(existingImage);
+        productImageRepository.deleteById(productImageId);
 
         return modelMapper.map(existingImage, ProductImageDto.class);
     }
